refactor : error of dto in swagger #11
