+++ src/main/java/com/example/umc8th/service/member/MemberCommandServiceImpl.java
@@ -1,13 +1,10 @@
 package com.example.umc8th.service.member;
 
 import com.example.umc8th.domain.member.Member;
-import com.example.umc8th.domain.member.MemberStatus;
-import com.example.umc8th.domain.member.Role;
 import com.example.umc8th.repository.member.MemberRepository;
 import com.example.umc8th.repository.mission.MyMissionRepository;
 import com.example.umc8th.repository.foodtype.PreferFoodTypeRepository;
-import com.example.umc8th.web.dto.converter.MemberConverter;
-import com.example.umc8th.web.dto.member.RequestMemberDto;
+import com.example.umc8th.web.converter.MemberConverter;
 import com.example.umc8th.web.dto.member.RequestMemberDto.JoinDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
