+++ src/main/java/com/example/umc8th/service/mission/MissionCommandService.java
@@ -4,7 +4,7 @@ import com.example.umc8th.web.dto.mission.RequestMissionDto;
 
 public interface MissionCommandService {
 
-    Long registerMission(Long restaurantId, RequestMissionDto.RegisterDto dto);
+    Long registerMission(Long restaurantId, RequestMissionDto.AddDto dto);
 
     Long addMissionOfMember(Long memberId, Long missionId);
 }
