+++ src/main/java/com/example/umc8th/web/converter/MemberConverter.java
@@ -1,4 +1,4 @@
-package com.example.umc8th.web.dto.converter;
+package com.example.umc8th.web.converter;
 
 import com.example.umc8th.domain.member.Member;
 import com.example.umc8th.domain.member.Role;
