+++ src/main/java/com/example/umc8th/web/converter/MissionConverter.java
@@ -1,11 +1,11 @@
-package com.example.umc8th.web.dto.converter;
+package com.example.umc8th.web.converter;
 
 import com.example.umc8th.domain.mission.Mission;
 import com.example.umc8th.web.dto.mission.RequestMissionDto;
 
 public class MissionConverter {
 
-    public static Mission registerToMissionEntity(RequestMissionDto.RegisterDto dto) {
+    public static Mission registerToMissionEntity(RequestMissionDto.AddDto dto) {
         return Mission.builder()
                 .point(dto.getPoint())
                 .content(dto.getContent())
