+++ src/main/java/com/example/umc8th/web/dto/mission/RequestMissionDto.java
@@ -1,7 +1,10 @@
 package com.example.umc8th.web.dto.mission;
 
+import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.extern.jackson.Jacksonized;
 
 import java.time.LocalDate;
 
@@ -9,7 +12,9 @@ import java.time.LocalDate;
 public class RequestMissionDto {
     @Data
     @Builder
-    public static class RegisterDto{
+    @NoArgsConstructor
+    @AllArgsConstructor
+    public static class AddDto{
         private String content;
         private LocalDate dueDate;
         private int point;
