+++ src/main/java/com/example/umc8th/web/dto/region/RequestRestaurantDto.java
@@ -1,7 +1,9 @@
 package com.example.umc8th.web.dto.region;
 
+import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 import lombok.extern.jackson.Jacksonized;
 
 
@@ -9,7 +11,8 @@ public class RequestRestaurantDto {
 
     @Data
     @Builder
-    @Jacksonized
+    @NoArgsConstructor
+    @AllArgsConstructor
     public static class RegisterDto{
         private String name;
         private String imgUrl;
