+++ src/main/resources/application.yml
@@ -17,3 +17,8 @@ spring:
         hibernate:
           SQL: DEBUG
 
+springdoc:
+  api-docs:
+    enabled: true
+  swagger-ui:
+    enabled: true
